Move config.py -> plex_trakt_sync/config.py
