+++ main.py
@@ -17,7 +17,7 @@ from json.decoder import JSONDecodeError
 
 import pytrakt_extensions
 from trakt_list_util import TraktListUtil
-from config import CONFIG
+from plex_trakt_sync.config import CONFIG
 
 import requests_cache
